pyproject.toml
[project]
name = "smart-campus-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.0",
    "google-generativeai>=0.8.5",
    "groq>=0.36.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "markdownify>=1.2.2",
    "pymupdf>=1.26.6",
    "python-pptx>=1.0.2",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.21",
]
